web/requirements.txt
amqp==2.3.2
authy==2.2.1
Babel==2.6.0
billiard==3.5.0.4
CacheControl==0.12.5
cachetools==2.1.0
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==2.1.2
django-appconf==1.0.2
django-celery-email==2.0.1
django-decorator-include==2.0
django-import-export==1.1.0
django-mail-templated==2.6.5
django-phonenumber-field==2.1.0
django-render-block==0.5
django-rest-swagger==2.2.0
django-twilio==0.9.3
django-uuidfield==0.5.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
fcm-django==0.2.19
firebase-admin==2.13.0
google-api-core==1.4.1
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
gunicorn==19.3.0
idna==2.6
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
msgpack==0.5.6
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.9
phonenumbers==8.8.4
protobuf==3.6.1
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pyfcm==1.4.5
PyJWT==1.6.4
PySocks==1.6.8
python-dotenv==0.7.1
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.18.4
requests-toolbelt==0.8.0
rsa==4.0
simplejson==3.16.0
six==1.11.0
tablib==0.12.1
twilio==6.20.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
xlrd==1.1.0
xlwt==1.3.0
